fix(container): respect include field when checking for Dockerfile
Previously, when using multiple container modules in one garden.yml,
it would be impossible to have one of them use a remote image if any
of the others had a Dockerfile in the same directory.
